Scraper/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
playwright
pydantic
httpx
pytest
requests
beautifulsoup4
selenium
webdriver-manager
lxml
fuzzywuzzy
python-levenshtein
tenacity